Add abstract job
### What changes are proposed in this pull request?

Add abstract job and delete copy implementation

### Why are the changes needed?

Please clarify why the changes are needed. For instance,
clean code

### Does this PR introduce any user facing changes?

na

pr-link: Alluxio/alluxio#17103
