+++ core/server/master/src/main/java/alluxio/master/job/JobFactoryProducer.java
@@ -11,7 +11,6 @@
 
 package alluxio.master.job;
 
-import alluxio.job.CopyJobRequest;
 import alluxio.job.JobRequest;
 import alluxio.job.LoadJobRequest;
 import alluxio.master.file.FileSystemMaster;
@@ -33,9 +32,6 @@ public class JobFactoryProducer {
     if (request instanceof LoadJobRequest) {
       return new LoadJobFactory((LoadJobRequest) request, fsMaster);
     }
-    if (request instanceof CopyJobRequest) {
-      return new CopyJobFactory((CopyJobRequest) request, fsMaster);
-    }
     throw new IllegalArgumentException("Unknown job type: " + request.getType());
   }
 
@@ -47,9 +43,6 @@ public class JobFactoryProducer {
   public static JobFactory create(Journal.JournalEntry entry, FileSystemMaster fsMaster) {
     if (entry.hasLoadJob()) {
       return new JournalLoadJobFactory(entry.getLoadJob(), fsMaster);
-    }
-    if (entry.hasCopyJob()) {
-      return new JournalCopyJobFactory(entry.getCopyJob(), fsMaster);
     }
     else {
       throw new IllegalArgumentException("Unknown job type: " + entry);
